Clean up samples
