+++ samples/CookieSessionSample/Startup.cs
@@ -51,16 +51,5 @@ namespace CookieSessionSample
                 await context.Response.WriteAsync("Hello old timer");
             });
         }
-
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .UseKestrel()
-                .UseIISIntegration()
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
     }
 }
