+++ samples/SocialSample/Properties/launchSettings.json
@@ -13,15 +13,16 @@
       "launchBrowser": true,
       "launchUrl": "https://localhost:44318/",
       "environmentVariables": {
-        "ASPNET_ENV": "Development"
+        "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "web": {
-      "commandName": "web",
+    "SocialSample": {
+      "commandName": "Project",
       "launchBrowser": true,
       "launchUrl": "https://localhost:54541/",
       "environmentVariables": {
-        "Hosting:Environment": "Development"
+        "ASPNETCORE_ENVIRONMENT": "Development",
+        "ASPNETCORE_SERVER.URLS": "https://localhost:54541/"
       }
     }
   }
